feat: add column def, remove zhtw, add default column

+++ app/Console/Commands/CreateTranslation.php
@@ -4,6 +4,8 @@ namespace App\Console\Commands;
 
 use App\Models\Translation;
 use Illuminate\Console\Command;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 use Storage;
 
 class CreateTranslation extends Command
@@ -24,17 +26,9 @@ class CreateTranslation extends Command
 
     protected $translations;
 
-    protected $table = [
-        'product' => [
-            'product_name',
-            'farm',
-        ],
-        'task_log' => [
-            'task',
-            'location',
-            'tool',
-            'tool_type',
-        ],
+    protected $tables = [
+        'product',
+        'task_log'
     ];
 
     /**
@@ -55,13 +49,29 @@ class CreateTranslation extends Command
      */
     public function handle()
     {
-
-        Translation::on('mysql_admin')->truncate();
-        foreach ($this->table as $table => $columns) {
+        if (!Schema::connection('mysql_admin')->hasTable('translations')) {
+            // 創建翻譯表
+            Schema::connection('mysql_admin')->create('translations', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('table_name', 50);
+                $table->string('column_name', 50);
+                $table->integer('foreign_key')->unsigned();
+                $table->string('locale', 10);
+                $table->text('value')->comment('翻譯');
+                $table->text('default')->comment('原始值');
+                $table->unique(['table_name', 'column_name', 'foreign_key', 'locale']);
+                $table->timestamps();
+            });
+        } else {
+            // 清空數據
+            Translation::on('mysql_admin')->truncate();
+        }
+        //產生資料
+        foreach ($this->tables as $table) {
             $model = studly_case($table);
-            $model = app("App\Models\\$model")->get();
-            foreach ($columns as $column) {
-                foreach ($model as $row) {
+            $model = app("App\Models\\$model");
+            foreach ($model->getTranslatableAttributes() as $column) {
+                foreach ($model->get() as $row) {
                     $this->insertRow($table, $column, $row->getKey(), $row->$column);
                 }
             }
@@ -71,16 +81,16 @@ class CreateTranslation extends Command
     protected function insertRow($table, $column, $foreign_key, $value)
     {
         $locales = [
-            'zhtw' => $value,
-            'en' => $this->translations[$value], //en
+            'en' => $this->translations[$value],
         ];
-        foreach ($locales as $locale => $value) {
+        foreach ($locales as $locale => $v) {
             Translation::on('mysql_admin')->create([
                 'table_name' => $table,
                 'column_name' => $column,
                 'foreign_key' => $foreign_key,
                 'locale' => $locale,
-                'value' => $value,
+                'value' => $v,
+                'default' => $value,
             ]);
         }
